MAINT: refactored code for fmri interfaces
